Restructure the project
- make the library header-only (removed the src folder and changed cmake)
- add unit tests (add tests folder and catch2 to cmake)
